+++ modules/portal/portal-background-task-test/test/unit/com/liferay/portal/background/task/internal/BaseBackgroundTaskTestCase.java
@@ -14,8 +14,8 @@
 
 package com.liferay.portal.background.task.internal;
 
-import com.liferay.portal.kernel.cache.Lifecycle;
-import com.liferay.portal.kernel.cache.ThreadLocalCacheManager;
+import com.liferay.portal.kernel.cache.thread.local.Lifecycle;
+import com.liferay.portal.kernel.cache.thread.local.ThreadLocalCacheManager;
 import com.liferay.portal.kernel.cluster.ClusterInvokeThreadLocal;
 import com.liferay.portal.kernel.util.CentralizedThreadLocal;
 import com.liferay.portal.kernel.util.GroupThreadLocal;
