+++ modules/portal/portal-cluster/src/com/liferay/portal/cluster/internal/ClusterRequestReceiver.java
@@ -16,8 +16,8 @@ package com.liferay.portal.cluster.internal;
 
 import com.liferay.portal.cluster.BaseClusterReceiver;
 import com.liferay.portal.cluster.ClusterChannel;
-import com.liferay.portal.kernel.cache.Lifecycle;
-import com.liferay.portal.kernel.cache.ThreadLocalCacheManager;
+import com.liferay.portal.kernel.cache.thread.local.Lifecycle;
+import com.liferay.portal.kernel.cache.thread.local.ThreadLocalCacheManager;
 import com.liferay.portal.kernel.cluster.Address;
 import com.liferay.portal.kernel.cluster.ClusterNode;
 import com.liferay.portal.kernel.cluster.ClusterNodeResponse;
