+++ portal-impl/src/com/liferay/portal/servlet/filters/threadlocal/ThreadLocalFilter.java
@@ -14,8 +14,8 @@
 
 package com.liferay.portal.servlet.filters.threadlocal;
 
-import com.liferay.portal.kernel.cache.Lifecycle;
-import com.liferay.portal.kernel.cache.ThreadLocalCacheManager;
+import com.liferay.portal.kernel.cache.thread.local.Lifecycle;
+import com.liferay.portal.kernel.cache.thread.local.ThreadLocalCacheManager;
 import com.liferay.portal.kernel.servlet.TryFinallyFilter;
 import com.liferay.portal.kernel.util.CentralizedThreadLocal;
 import com.liferay.portal.servlet.filters.BasePortalFilter;
