+++ portal-impl/src/com/liferay/portal/xml/xpath/MatchesFunction.java
@@ -14,9 +14,9 @@
 
 package com.liferay.portal.xml.xpath;
 
-import com.liferay.portal.kernel.cache.Lifecycle;
-import com.liferay.portal.kernel.cache.ThreadLocalCache;
-import com.liferay.portal.kernel.cache.ThreadLocalCacheManager;
+import com.liferay.portal.kernel.cache.thread.local.Lifecycle;
+import com.liferay.portal.kernel.cache.thread.local.ThreadLocalCache;
+import com.liferay.portal.kernel.cache.thread.local.ThreadLocalCacheManager;
 
 import java.util.HashMap;
 import java.util.List;
